requirements.txt
streamlit
dash
plotly
pandas
numpy
jupytext
openpyxl
optuna
pandas_gbq
scikit-learn
lightgbm
japanize-matplotlib
networkx
dtreeviz
dash-cytoscape
pltvid 
rfpimp